# apps/web/app/api/auth/login/route.ts
import { NextResponse } from "next/server";
import { LoginBody, TokenPair, ErrorCodes } from "@taakhelden/shared";
import { apiFetch, crossOriginBlock, forwardHeaders } from "../../../../lib/api/config";
import { setTokens } from "../../../../lib/auth/session";

/** BFF login: forwards credentials to the Worker and stores tokens in httpOnly cookies. */
export async function POST(req: Request) {
  const blocked = crossOriginBlock(req);
  if (blocked) return blocked;

  const parsed = LoginBody.safeParse(await req.json().catch(() => null));
  if (!parsed.success) {
    return NextResponse.json(
      { error: { code: ErrorCodes.VALIDATION_FAILED, message: "Ongeldige invoer." } },
      { status: 400 },
    );
  }

  let res: Response;
  try {
    res = await apiFetch("/auth/login", {
      method: "POST",
      headers: forwardHeaders(req, { "Content-Type": "application/json" }),
      body: JSON.stringify(parsed.data),
      cache: "no-store",
    });
  } catch {
    return NextResponse.json(
      { error: { code: ErrorCodes.UPSTREAM_UNAVAILABLE, message: "Kan de server niet bereiken." } },
      { status: 502 },
    );
  }

  const data = await res.json().catch(() => null);

  if (!res.ok) {
    if (!data) {
      return NextResponse.json(
        { error: { code: ErrorCodes.UPSTREAM_UNAVAILABLE, message: "Kan de server niet bereiken." } },
        { status: 502 },
      );
    }
    return NextResponse.json(data, { status: res.status });
  }

  const tokens = TokenPair.safeParse(data);
  if (!tokens.success) {
    return NextResponse.json(
      { error: { code: ErrorCodes.VALIDATION_FAILED, message: "Onverwacht antwoord van de server." } },
      { status: 502 },
    );
  }

  await setTokens(tokens.data);
  return NextResponse.json({ ok: true });
}