# docs/openapi/taakhelden-core-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wispel core API",
    "version": "0.1.0",
    "description": "Generated core contract snapshot from packages/shared. Covers web dashboard and iOS Phase 1 endpoints."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Log in a parent account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentSessionResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/photos/upload-intent": {
      "post": {
        "summary": "Create a photo upload intent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadIntentBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadIntentResponse"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/photos/{id}/confirm": {
      "post": {
        "summary": "Confirm an uploaded photo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoConfirmResponse"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/photos/{id}": {
      "get": {
        "summary": "Read photo status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "$ref": "#/definitions/ApiError",
        "definitions": {
          "ApiError": {
            "type": "object",
            "properties": {
              "error": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "enum": [
                      "VALIDATION_FAILED",
                      "UNAUTHORIZED",
                      "FORBIDDEN",
                      "NOT_FOUND",
                      "TASK_ALREADY_COMPLETED",
                      "INVALID_STATUS",
                      "UNDO_WINDOW_EXPIRED",
                      "EMAIL_IN_USE",
                      "INVALID_CREDENTIALS",
                      "INSUFFICIENT_POINTS",
                      "REWARD_LIMIT_REACHED",
                      "PIN_LOCKED",
                      "INVALID_FAMILY_CODE",
                      "INVALID_INVITE",
                      "INSTANCE_SLOT_TAKEN",
                      "IDEMPOTENCY_REPLAY",
                      "IDEMPOTENCY_KEY_REUSED",
                      "RATE_LIMITED",
                      "UPSTREAM_UNAVAILABLE"
                    ]
                  },
                  "message": {
                    "type": "string"
                  },
                  "details": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "code",
                  "message"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "error"
            ],
            "additionalProperties": false
          }
        }
      },
...
      "LoginBody": {
        "$ref": "#/definitions/LoginBody",
        "definitions": {
          "LoginBody": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email"
              },
              "password": {
                "type": "string"
              }
            },
            "required": [
              "email",
              "password"
            ],
            "additionalProperties": false
          }
        }
      },
...
      "ParentSessionResult": {
        "$ref": "#/definitions/ParentSessionResult",
        "definitions": {
          "ParentSessionResult": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              },
              "expiresIn": {
                "type": "number"
              },
              "userId": {
                "type": "string"
              },
              "familyId": {
                "type": "string"
              }
            },
            "required": [
              "accessToken",
              "refreshToken",
              "expiresIn",
              "userId",
              "familyId"
            ],
            "additionalProperties": false
          }
        }
      },
...
      "UploadIntentBody": {
        "$ref": "#/definitions/UploadIntentBody",
        "definitions": {
          "UploadIntentBody": {
            "type": "object",
            "properties": {
              "purpose": {
                "type": "string",
                "enum": [
                  "task",
                  "profile"
                ]
              },
              "instanceId": {
                "type": "string"
              },
              "memberId": {
                "type": "string"
              },
              "contentType": {
                "type": "string",
                "enum": [
                  "image/jpeg",
                  "image/heic",
                  "image/png"
                ]
              },
              "bytes": {
                "type": "integer",
                "minimum": 1,
                "maximum": 10485760
              }
            },
            "required": [
              "purpose",
              "contentType",
              "bytes"
            ],
            "additionalProperties": false
          }
        }
      },
      "UploadIntentResponse": {
        "$ref": "#/definitions/UploadIntentResponse",
        "definitions": {
          "UploadIntentResponse": {
            "type": "object",
            "properties": {
              "photoId": {
                "type": "string"
              },
              "uploadUrl": {
                "type": "string",
                "format": "uri"
              }
            },
            "required": [
              "photoId",
              "uploadUrl"
            ],
            "additionalProperties": false
          }
        }
      },
      "PhotoStatusResponse": {
        "$ref": "#/definitions/PhotoStatusResponse",
        "definitions": {
          "PhotoStatusResponse": {
            "type": "object",
            "properties": {
              "photoId": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "intent",
                  "uploaded",
                  "processing",
                  "ready",
                  "failed"
                ]
              },
              "url": {
                "type": "string",
                "format": "uri",
                "nullable": true
              }
            },
            "required": [
              "photoId",
              "status",
              "url"
            ],
            "additionalProperties": false
          }
        }
      },
      "PhotoConfirmResponse": {
        "$ref": "#/definitions/PhotoConfirmResponse",
        "definitions": {
          "PhotoConfirmResponse": {
            "type": "object",
            "properties": {
              "photoId": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "intent",
                  "uploaded",
                  "processing",
                  "ready",
                  "failed"
                ]
              }
            },
            "required": [
              "photoId",
              "status"
            ],
            "additionalProperties": false
          }
        }
      },
...
    }
  }
}